+++ app/src/main/java/com/cs2/veterinaria/app/domains/model/Bill.java
@@ -9,9 +9,9 @@ import lombok.Setter;
 @NoArgsConstructor
 public class Bill {
     private Long idBill;
-    private int idOrder;
+    private long idOrder;
     private long idPet;
     private long idOwner;
     private String dateCreated;
-    private float price;
+    private double price;
 }
