+++ app/src/main/java/com/cs2/veterinaria/app/domains/model/Order.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Getter
 @NoArgsConstructor
 public class Order {
-    private int idOrder;
+    private long idOrder;
     private long idPet;
     private long idOwner;
     private long idVerterinarian;
